fix(P5-02): MinSupportedSampleRate is a Double (Duration), not UInt32
OPC UA Part 5 §6.3.2: MinSupportedSampleRate is a Duration (Double). The core node
manager exposed its Value as `min_sampling_interval_ms as u32`, so a client read got
UInt32(100) while the node's DataType attribute is Duration (NodeId i=290) — a
datatype mismatch (verified at runtime). The value is already an f64; exposed
directly via .into() (the old cast also truncated the Double).

Resolves the earlier multi-AI conflict in AG's favour (the Claude agent's HONORED was
wrong). Test min_supported_sample_rate_value_matches_duration_datatype (red-first,
reads Value + DataType). Server lib 136 green; clippy --all-targets + fmt clean.

+++ async-opcua-server/src/node_manager/memory/core.rs
@@ -327,7 +327,9 @@ impl CoreNodeManagerImpl {
                 (limits.max_string_length as u32).into()
             }
             VariableId::Server_ServerCapabilities_MinSupportedSampleRate => {
-                (limits.subscriptions.min_sampling_interval_ms as u32).into()
+                // Part 5 §6.3.2: MinSupportedSampleRate is a Duration (Double), not a UInt32 — the
+                // value is already an f64, so expose it directly (the old `as u32` truncated it).
+                limits.subscriptions.min_sampling_interval_ms.into()
             }
             VariableId::Server_ServerCapabilities_OperationLimits_MaxMonitoredItemsPerCall => {
                 (limits.operational.max_monitored_items_per_call as u32).into()
